refactor(cl): convert SitemapTest to mixin
