+++ cl/opinion_page/tests.py
@@ -30,7 +30,6 @@ from cl.citations.utils import slugify_reporter
 from cl.lib.models import THUMBNAIL_STATUSES
 from cl.lib.redis_utils import get_redis_interface
 from cl.lib.storage import clobbering_get_name
-from cl.lib.test_helpers import SitemapTest
 from cl.opinion_page.forms import (
     MeCourtUploadForm,
     MissCourtUploadForm,
@@ -89,6 +88,7 @@ from cl.tests.mixins import (
     PeopleMixin,
     SearchMixin,
     SimpleUserDataMixin,
+    SitemapMixin,
 )
 from cl.tests.providers import fake
 from cl.users.factories import UserFactory, UserProfileWithParentsFactory
@@ -972,9 +972,10 @@ class NewDocketAlertTest(SimpleUserDataMixin, TestCase):
         self.assertInHTML("Get Docket Alerts", r.content.decode())
 
 
-class OpinionSitemapTest(SitemapTest):
+class OpinionSitemapTest(SitemapMixin, TestCase):
     @classmethod
     def setUpTestData(cls) -> None:
+        super().setUpTestData()
         # Included b/c so new
         OpinionClusterWithParentsFactory.create(
             precedential_status=PRECEDENTIAL_STATUS.PUBLISHED,
@@ -1001,6 +1002,7 @@ class OpinionSitemapTest(SitemapTest):
         )
 
     def setUp(self) -> None:
+        super().setUp()
         self.sitemap_url = reverse(
             "sitemaps", kwargs={"section": SEARCH_TYPES.OPINION}
         )
@@ -1010,9 +1012,10 @@ class OpinionSitemapTest(SitemapTest):
         super().assert_sitemap_has_content()
 
 
-class DocketSitemapTest(SitemapTest):
+class DocketSitemapTest(SitemapMixin, TestCase):
     @classmethod
     def setUpTestData(cls) -> None:
+        super().setUpTestData()
         # Included b/c so new
         DocketFactory.create(
             source=Docket.RECAP,
@@ -1034,6 +1037,7 @@ class DocketSitemapTest(SitemapTest):
         )
 
     def setUp(self) -> None:
+        super().setUp()
         self.setUpSiteDomain()
 
         self.sitemap_url = reverse(
@@ -1064,9 +1068,10 @@ class DocketSitemapTest(SitemapTest):
         super().assert_sitemap_has_content()
 
 
-class DocketEmptySitemapTest(SitemapTest):
+class DocketEmptySitemapTest(SitemapMixin, TestCase):
     @classmethod
     def setUpTestData(cls) -> None:
+        super().setUpTestData()
         # Excluded b/c old
         DocketFactory.create(
             source=Docket.RECAP,
@@ -1087,6 +1092,7 @@ class DocketEmptySitemapTest(SitemapTest):
         )
 
     def setUp(self) -> None:
+        super().setUp()
         self.setUpSiteDomain()
 
         self.sitemap_url = reverse(
@@ -1112,11 +1118,12 @@ class DocketEmptySitemapTest(SitemapTest):
         )
 
 
-class BlockedSitemapTest(SitemapTest):
+class BlockedSitemapTest(SitemapMixin, TestCase):
     """Do we create sitemaps of recently blocked opinions?"""
 
     @classmethod
     def setUpTestData(cls) -> None:
+        super().setUpTestData()
         # Included b/c recently blocked
         OpinionClusterWithParentsFactory.create(
             blocked=True,
@@ -1129,6 +1136,7 @@ class BlockedSitemapTest(SitemapTest):
         )
 
     def setUp(self) -> None:
+        super.setUp()
         self.sitemap_url = reverse(
             "sitemaps", kwargs={"section": "blocked-opinions"}
         )
